Update interface.py
updated spairrowtuning. i dont know if the ret.safetyparam will work because i think it needs to be changed in the dbc (please advise). also geo said steerup15 is not banned so if you can make a parameter in op edit then it will make my tune complete. i hope to use your fork to provide a few individuals routes for the PR to master.