Update toyota_nodsu_pt_generated.dbc
this would apply to all toyotas but i dont know how to make a separate dbc file to access for just one OP EDIT param. set to corolla_use_indi and it allows a 53 safetyparam? for a safety param to apply it requires a dbc edit and a interface edit, can corolla tss2 have its own dbc file on your fork?