Enable Toyota Button on Steering Wheel to Control Dynamic Follow
This adds a new op_param: `toyota_distance_btn`

When set to `True` this will enable the distance button on
Toyota steering wheels to control the dynamic follow profile.

This has a couple caveats:
- This is limited to TSS2 Toyotas ONLY at the moment.  I need
cabana data from non-TSS2 cars to fix this.  There is some
difference in the messaging that is causing a test failure
on car models that are not TSS2.
- The df button on the screen is removed when enabled.
- There is no way to enable the `auto` profile this way
- The profile always resets to `stock` when the car is started,
this is a result of Toyota's defaults.