Add Support for Distance Button Control on Toyota TSS1 with SDSU
- Updates the opendbc submodule to the current version with additional
commits containing support for the SDSU FD_BUTTON message.

- Updates the Cereal Submodule to a version with hasSmartDSU support.

If toyota_distance_btn is enabled and the vehicle has an SDSU,
forward the FD_BUTTON bit to the vehicle to control the distance
lines.

Note this will require that the SDSU has at least this firmware
on it, in order for the ditance button signal to be forwarded:
https://github.com/wocsor/panda/commit/b5120f6427551345c543b490fe47da189c1e48e1